refactor(frontend): 命名・型の厳密化（ViewKind/DeleteFileAction/LibrarySettingsView）
SettingsView.svelteをLibrarySettingsView.svelteにリネーム（App.svelteの
import元を更新）。App.svelteのcurrentViewリテラルユニオンをtypes.tsの
ViewKindとしてエクスポートし、Sidebar.svelte/LibrarySettingsView.svelteの
onNavigateのstring受けとApp.svelteのasキャストを排除。WorkGrid.svelteの
resolveFileActionをresolveEffectiveDeleteActionにリネームし、
selectedFileActionの型をDeleteFileAction | "none"に厳密化。App.svelteの
initTheme/initBannerAutoCloseをonMountに移動（loadActiveLibraryは
$effectのまま非対称に維持）。Sidebar.svelteのtagsByCategory計算を
src/lib/utils/tagGrouping.tsのプレーン関数へ切り出し、
svelte/prefer-svelte-reactivityのeslint-disableを解消。
